[refactor] 패키지 명 컨벤션에 맞게 변경
