+++ src/test/java/team7/inplace/video/persistence/VideoRepositoryTest.java
@@ -1,4 +1,4 @@
-package team7.inplace.video.repository;
+package team7.inplace.video.persistence;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
@@ -15,7 +15,6 @@ import org.springframework.test.annotation.DirtiesContext;
 import team7.inplace.influencer.domain.Influencer;
 import team7.inplace.place.domain.*;
 import team7.inplace.video.domain.Video;
-import team7.inplace.video.persistence.VideoRepository;
 
 @DataJpaTest
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD) // 각 메서드 실행마다 이전 결과 초기화
